change a file blue print
