+++ models.py
@@ -6,7 +6,6 @@ datetime.utcnow()
 
 db=SQLAlchemy()
 
-
 class User(db.Model):
     id = db.Column(db.Integer, primary_key=True)
     name = db.Column(db.String(200), nullable=False)
